+++ test/chrome/windows.main.ts
@@ -1,3 +1,3 @@
 import Windows from '../../src/browser/handlers/windows';
 
-new Windows('test', console.log);
+new Windows('test', () => { });
